fixed bug when phi=0

+++ lib/common.py
@@ -367,7 +367,7 @@ def weighted_order_choose(orders, n, feekey):
 	if phi > 0:
 		weight = np.exp(-(1.0*fee - minfee) / phi)
 	else:
-		weight = np.ones_like(fee)
+		weight = np.ones_like(fee, dtype=np.float)
 	weight /= sum(weight)
 	debug('randomly choosing orders with weighting\n' + pprint.pformat(zip(orders, weight)))
 	chosen_order_index = np.random.choice(len(orders), p=weight)
